More Menu assets and special world assets
Kind of screwed around on a lot of different things trying to get them to work together. Like Player dying or transitions between worlds and the secret world. Also added a new clause for the Background music to stop playing it properly. Fixed a few bugs with firefly and falling rock and ToxicWater. Added a Credits script to the Credits scene.